devlopment phase 1.1.1

+++ src/main/java/com/exaple/quiet_Q/modal/Community.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class Community {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
 
     private String communityName;
 
@@ -23,18 +23,18 @@ public class Community {
     public Community() {
     }
 
-    public Community(Integer id, String communityName, List<User> users, List<Post> posts) {
+    public Community(Long id, String communityName, List<User> users, List<Post> posts) {
         this.id = id;
         this.communityName = communityName;
         this.users = users;
         this.posts = posts;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
